Tadpole: Update to version 1.4
New:
- support for conditional variables (not just conditional blocks)

Support for conditional variables was added to the existing framework for conditional
blocks, namely the [tp-if:] tag. A conditional variable is assumed when a [tp-if] tag
is encountered which does not contain a ';block=' modifier.

- enhanced alternative text options

Enhanced support for ';empty=' modifier. Now alternative text can be surrounded by backwards
single quote (`) marks. This allows for both single (') and double (") quotation marks
to be used freely in the alternative content. Note that the only character which can NOT
be used in alternative content blocks is the semicolon (;) as that is interpreted as the
modifier separator. Future releases may adjust this. 

- enhanced conditional logic

Both conditional blocks and variables utilize a more flexible
logic for determining the condition in which they validate. The right-hand side of the
equation may now also be a variable, either absolute (tp.*) or relative (@.*).