modified FRouter::Route to accept a base url (default = '/') to prepend to each request. This allows the application to live anywhere in the directory structure (instead of requiring that it reside at the server root).
The Entrance class passes the configuration variable 'fconfig_url_base' to
the router, ensuring that whichever value the user has specified in the 
configuration file will be prepended to each request. This avoids having to 
define special routes, and makes the application much more portable. As an example:
to go from a debug environment in a subdirectory to a production environment at the 
server root, all that is required is to change the value of 'fconfig_url_base' from
'/some/sub/directory' to '/'