Introduced formal support for "boolean" attributes.
When an attribute is specified as boolean, its value is either "true" (1)
or "false" (0). The value can be accessed by prepending the word "is" to 
the variable name. Examples:

Say you have a Person object with a boolean variable "living" describing whether
or not the individual is alive.
$p = new Person();
To test whether the person is alive:
($p->isAlive()) ? "Alive" : "Not Alive"

By contrast, all other (non-boolean) local variables (age, for instance) are
accessed by prepending the word "get" to the variable name. Example:
$age = $p->getAge();

Finally, note that "setting" a boolean variable is accomplished the same way
any other variable is set, namely with a call to set + variable name. Example:
$p->setAlive(true);
$p->setAge(39);