finally resolve issue with save/update calls not persisting all of the information.
In PHP, isset() returns true only if the value exists and is not null. in the case of a new object, all of the attributes are null and so isset always fails, meaning no data gets merged in. Using array_key_exists() on the $properties array (result of object_get_vars()) tests whether a matching attribute has been defined, regardless of its value, which is the desired behavior.