Furnace: Security update
Support password salts and centralize password encryption algorithm into FAccountManager::EncryptPassword(). This static function should be called any time an encrypted password needs to be generated from plaintext.

Application config file (/app/config/app.yml) changes:

password_salt: <random string of characters>

If it is missing, passwords will not be salted and will be hashed as-is. If it is added, accounts created *before* adding it will no longer be able to log in with their old passwords.