add a provider folder for frameworkers.org and put a sample helper class inside. The MenuListHelper class generates an html unordered list from a specially crafted array of menu items:
$menuItems = array(
	array("item1Label","/item1/href","item1 tooltip description",array(..item1 sub items..))
	...
);

will become:

<ul>
  <li><a href="/item1/href" title="item1 tooltip description">item1Label</a></li>
  ...
</ul>

If an item has sub-items, those will be expanded into a nested <ul/> object. The helper even does rudimentary matching on the "href" attributes to determine which (if any) should be considered the "active" element. If one is active, its <li> has the class "active": <li class="active".../>