changes to the way javascript and css are referenced from within page layouts:
instead of the old cumbersome way of specifying a path and complicated tadpole tag, now just place:

[_js_]   and   [_css_]

anywhere in the <head> section of your layouts, and the included javascript and css files will be placed there.

Also, new ability to 'inject' javascript and css style into the page from the controller. This can be very useful when dynamism is required. From any controller, inject arbitrary js or css by using $this->injectJS() or $this->injectCSS() and it will be placed inside the <head> section of the page when it is rendered