FModel: when an object attribute has defined a set of 'allowedValues' in its validator, create a corresponding set of static variables (one per allowed value) and a corresponding decoder function. This allows developers to minimize the number of "magic numbers" that get sprinkled through the code.
All generated variables will be of the form <var>_<label>, so if the 'allowedValues' section of the Transaction::type attribute has the following values:
0|Income
1|Expense
2|Transfer

Then the following static variables will be available in the Transaction class:

public static $TYPE_INCOME = 0;
public static $TYPE_EXPENSE= 1;
public static $TYPE_TRANSFER = 2;

This means that, instead of this:

$transaction->setType(2);

developers can now do this: 

$transaction->setType(Transaction::$TYPE_EXPENSE);


The decoder provides the opposite translation, returning the specified human-readable value given the code. It is static so that it can be called from anywhere.

public static function DecodeType($value) {
  switch ($value) {
    case self::$TYPE_INCOME: return "Income";
    case self::$TYPE_EXPENSE: return "Expense";
    case self::$TYPE_TRANSFER: return "Transfer";
    default: return "Unknown";
  }
}

Because all of this is now built into the framework, there is no collateral damage when an allowed value is changed via Fuel... everything is simply re-generated with the updated value.