Added a 'decode' command. This takes advantage of the newly built-in 'allowedValue' decoder functionality. Here's an example:
Transaction::type is defined as an integer, and has three allowed values:
0|Income
1|Expense
2|Transfer

It will most often be the case that the 'human-readable' version of Transaction::type (rather than the coded integer value) is desired for display in a view. Whenever this is the case, developers have only to add the 'decode' command to the appropriate tag, like so:

[transaction.type;decode]

Without the command, the output for an 'expense' transaction would be:

[transaction.type]  ==> 1

With the command, the output would be:

[transaction.type;decode] ==> 'Expense'

This removes the need to write 'getHumanReadableXXX' type functions, and, since everything integrated at the framework level, the values are always kept up-to-date when changes are made in Fuel.