Add FPage::setJs(). This behaves exactly like FPage::set() but instead of passing a value to the view manager (for use in Tadpole tags), it injects the variable into a <script> block in the page head (for use by javascripts).
For example, if a javascript file needed the username of the currently logged in user, one could do the following, from within the controller:

$this->setJs('username',_user()->getUsername());  // 'andrew';

which would cause:

<script type="text/javascript">
  var username = 'andrew';
</script>

to get injected into the <head> tag of the resulting page. Since it is global in scope, any javascript run from any file can now reference 'username'.