Added support for a new Tadpole extension: Edit-In-Place.
Tadpole has been built in a way that makes it easy to add functionality. This commit defines a new type of Tadpole tag, the 'edit-in-place' tag, which when rendered provides an AJAX-driven edit-in-place input element much like is commonly seen on Flickr and other "Web-2.0" sites.

Tag structure:

[eip:object.attribute;context="contextString"]

Optional Commands:

;action='/url/to/handler] : specify a custom handler for this element. The default utilizes the Frameworkers Ajax extension.

;callback='callbackFunction' : specify a javascript function to invoke with the response from the server. See 'Dependencies' for more details on the nature of the response

;inline : If present, the form will be rendered inline rather than as a block, allowing for neater presentation in some cases

Dependencies:

- Frameworkers Ajax extension: This ships standard with every Furnace installation, so it should already be present
- jQuery: (http://jquery.com) The EIP feature is currently very tied to jQuery. Maybe it can be refactored in the future, but at the moment it is a firm requirement. 
- jEditable plugin for jQuery: (http://www.appelsiini.net/projects/jeditable) This script provides the marshalling of data and options between client and server. Just make sure this script is available inside the 'js' directory of your current theme.

Usage:

To use the EIP tag, you need an object and a security context string. For this example, we will assume:
- A Person (USER) has many to-do items (TASKs) that belong to named lists (LIST)

To edit the description "in-place" for task 't' with id=7, use the following tag:
Task Description: [eip:t.description;context="tasks{7}"]

The security context is a string that allows evaluation, starting at _user(), and ending with a single data object. It is this object whose attributes will be updated by the edit. Because evaluation starts with _user() (the currently logged in user), this mechanism ensures that no user will ever be able to make edits to data that s/he does not already have access to. In the example above, the context gets evaluated to: _user()->get('tasks')->unique('id',7). If the user does not have access to task 7, the above will fail, and the user will see a failure (not authorized) notice. More complicated contexts are possible, see the documentation in the code (updateController.php in the Frameworkers Ajax extension) for a more technical description.