1) change (simplify) the way important paths are represented in Tadpole tags.
old        -->    new
[_theme_]         [%theme]
[_local_]         [%local]
[_app.url_base]   [%a]

The [%a] tag is particularly useful when designing sites that may live outside the server root (i.e.: in some subdirectory). The [%a] tag will always expand to the value of 'url_base' in the application config file. 

Example: 
By using [%a] before all links sitewide, it becomes trivial to move the entire site from the server root (http://somesite.com/) to a subdirectory (http://somesite.com/mysite/) because, in the first case, [%a] = '' and in the second [%a] = '/mysite', so no code changes are necessary if all <a> tags use the [%a] tag like so:

<a href="[%a]/some/page">Some Page</a>

2) drop the requirement that each view
*must* exist within its own folder of the 
same name. The framework will now first check for an .html file matching the requested view, and if none exists, will check for a directory matching the requested view.
This means that it is only necessary to create a folder for a view if there are view-local assets (javascripts,css, etc) pertaining to the view.

normal view with no local static assets:
/pages/test/normalView.html

view with local static assets: (needs subfolder)
/pages/test/extraView/extraView.html
/pages/test/extraView/extra.css