implement FPage::setTitle to allow pages to specify a custom title
The default value for the title is taken from the 'name' attribute of the application configuration file.

The argument $append is a boolean flag (default = false) controlling whether the specified value is appended to the existing title or not. 

Example:

Say 'name' is 'MyApp' in config/app.yml, then:

$this->setTitle('foo') would generate:
<title>foo</title>

and 

$this->setTitle('foo',true) would generate:
<title>MyApp - foo</title>